refactor: rename frontend/ → app/ and Frontend struct → App
Module name now matches its role (the App runner / dispatcher), and
the struct name matches the module — consistent triple. Mechanical
rename, no behaviour change.

Mechanics:

- `git mv src/frontend src/app`
- `lib.rs`: `pub mod frontend;` → `pub mod app;`
- Struct `Frontend` → `App` (mass `\bFrontend\b` rename)
- Mass `crate::frontend` → `crate::app`, `ttymap::frontend` → `ttymap::app`
- Path-style doc references `` `frontend/` `` → `` `app/` ``,
  `` `frontend::` `` → `` `app::` `` updated across `src/`, `docs/`,
  `CLAUDE.md`, `README.md`
- `let mut frontend = ...` local in `main.rs` → `let mut app = ...`
- Plain-prose "the frontend layer" in module docstrings kept as-is
  (still meaningful as a generic term)

After this PR:

```
src/
├── app/             ← runner + dispatch + UserIntent + AppEvent + ui shell
│   ├── mod.rs       (App struct, run loop, dispatch)
│   ├── intent.rs    UserIntent
│   ├── event.rs     AppEvent
│   ├── ui.rs
│   └── frame_timer.rs
├── compositor/
├── palette/
├── input/
├── lua/
├── map/
├── theme/, geo/, shared/, commands/, config.rs, logging.rs
```

388 / 388 tests pass; clippy / fmt clean.
