+++ CLAUDE.md
@@ -23,8 +23,8 @@ The build step compiles `proto/vector_tile.proto` using protox (no system protoc
 ## Design philosophy
 
 See [docs/design.md](docs/design.md) for load-bearing design decisions:
-- **When to emit a `UserIntent` vs a direct method call** — user intent goes through `Frontend::dispatch`; internal data flow (frame arrival, widget polling) does not.
-- **Controller split: by feature, not by domain** — if `Frontend::dispatch` + cross-cutting helpers grow large.
+- **When to emit a `UserIntent` vs a direct method call** — user intent goes through `App::dispatch`; internal data flow (frame arrival, widget polling) does not.
+- **Controller split: by feature, not by domain** — if `App::dispatch` + cross-cutting helpers grow large.
 - **Cleanup via `Drop`, not manual** — `RenderHandle`'s thread shutdown is handled by its Drop impl.
 - **Frames are completed products** — main thread displays, does not compute.
 
@@ -34,7 +34,7 @@ For the full system architecture (src tree, layering, message + render flow, foc
 
 The app uses a **three-thread model**:
 
-1. **Main thread** (`src/frontend/`): Runs the event loop — drains completed frames from the render thread, polls plugins for async work, processes keyboard/mouse/resize events via crossterm, and asks ratatui to paint. State changes driven by user intent flow through `Frontend::dispatch` (the single Receiver), which speaks the `UserIntent` vocabulary defined in `src/frontend/intent.rs`.
+1. **Main thread** (`src/app/`): Runs the event loop — drains completed frames from the render thread, polls plugins for async work, processes keyboard/mouse/resize events via crossterm, and asks ratatui to paint. State changes driven by user intent flow through `App::dispatch` (the single Receiver), which speaks the `UserIntent` vocabulary defined in `src/app/intent.rs`.
 
 2. **Render thread** (`src/map/render/thread.rs`): Owns a `RenderPipeline` (tile cache + renderer). Receives `RenderTask` messages (`Draw(Viewport)` / `Resize` / `SetStyler` / `Shutdown`) via crossbeam-channel, and sends completed `MapFrame`s back. The loop is **purely event-driven**: a `crossbeam::select!` parks on the task channel and on a wake channel pinged by the decoder thread on each tile arrival — no timeout-based polling.
 
@@ -55,9 +55,9 @@ Key modules:
 - **`src/map/tile/disk.rs`**: Free-function disk read/write helpers used by both `fetch::DiskCachedFetcher` (worker-side, read+write through) and `cache::TileCache` (render-thread fast path, read-only). Layout: `{cache_dir}/{z}/{x}-{y}.pbf`.
 - **`src/map/tile/fetch/`**: `TileFetcher` (per-backend trait — "key → bytes"), the generic `FetchLane<F>` (queue / workers / dedup / priority), and the `DiskCachedFetcher<F>` decorator that adds a disk read-through / write-through layer to any inner fetcher. `http.rs` is the only inner backend today; new backends (mbtiles, pmtiles, …) add a `TileFetcher` impl + a branch in `app::build_tile_cache`.
 - **`src/map/tile/decoder.rs`**: Single-thread relay that reads bytes from the fetch lane, calls `decode::decode`, and forwards `DecodedTile`s to the cache. Empty bytes (negative cache from failed fetches) bypass `decode()` and surface as `DecodedTile::empty()`.
-- **`src/frontend/`**: `Frontend` is the sole Receiver and owns the latest `MapFrame` directly (no `UiState` wrapper — built-in chrome lives in plugins now). `src/frontend/intent.rs` holds the `UserIntent` intent vocabulary (map-level actions nest under `UserIntent::Map(MapAction)` because `MapState` owns its own vocabulary; other variants sit at the top level); `src/frontend/event.rs` holds `AppEvent`, the unified queue payload (`Intent` / `FrameReady` / `Input` / `Wake`); `src/frontend/mod.rs` holds `Frontend::dispatch` (the thin router) and cross-cutting methods like `switch_theme` / `handle_resize`. UI infrastructure lives under `frontend/`: `compositor/` (focus/modal stack), `palette/` (`:`-triggered picker), `ui.rs` (ratatui draw entry), `frame_timer.rs` (per-iteration wake source). Single side-effect boundary for user-intent state changes. See [docs/design.md](docs/design.md) for the UserIntent-vs-direct-API judgment rules.
+- **`src/app/`**: `App` is the sole Receiver and owns the latest `MapFrame` directly (no `UiState` wrapper — built-in chrome lives in plugins now). `src/app/intent.rs` holds the `UserIntent` intent vocabulary (map-level actions nest under `UserIntent::Map(MapAction)` because `MapState` owns its own vocabulary; other variants sit at the top level); `src/app/event.rs` holds `AppEvent`, the unified queue payload (`Intent` / `FrameReady` / `Input` / `Wake`); `src/app/mod.rs` holds `App::dispatch` (the thin router) and cross-cutting methods like `switch_theme` / `handle_resize`. UI infrastructure lives under `app/`: `compositor/` (focus/modal stack), `palette/` (`:`-triggered picker), `ui.rs` (ratatui draw entry), `frame_timer.rs` (per-iteration wake source). Single side-effect boundary for user-intent state changes. See [docs/design.md](docs/design.md) for the UserIntent-vs-direct-API judgment rules.
 - **`src/compositor/`**: Stack-based focus/modal system (helix-inspired). One primitive: a stack of `Component`s, where the top owns key focus. Components render side panels through `Component::render` and can emit `UserIntent` via `Window`. World-space overlays (markers etc.) are *not* a Component concern — every Lua plugin's per-frame map paint runs through the unified `LuaTickRegistry::tick` (called from `ui::draw`) which hands the plugin a `MapApi` it draws into directly. Submodule `map_api.rs` (the drawing facade) is a crate-internal type consumed by the Lua bridge. The earlier `src/plugin_api/` directory that hosted these helpers was a misnamed bucket from the pre-Lua era; it's gone now that every plugin author goes through Lua. `Placement` has two variants: `Floating` (palette-only, drawn over the map) and `Sidebar` (left rail, equal-vertical-split among up to 3 visible cards). Lua plugins always land in `Sidebar`.
-- **`src/frontend/ui.rs`**: Single thin `draw()` entry — no state of its own. Lays out the map area + footer and routes through the compositor for overlays and panels.
+- **`src/app/ui.rs`**: Single thin `draw()` entry — no state of its own. Lays out the map area + footer and routes through the compositor for overlays and panels.
 - **`src/palette/`**: `:`-triggered command palette as an ephemeral `Component` (state per-open, discarded on pop). Provider sub-modes (theme picker, forward-geocode search) swap in place via `PaletteAction::SwitchProvider` or are pushed pre-loaded by their key activation. Providers can be sync (`OnEachKey` filter — command, theme) or async (`Debounced` filter, `poll()` to drain results, `is_loading()` for the spinner — search).
 - **`src/commands/`**: CLI subcommands (currently `snap`). Each subcommand is one file with a `run()` entry point.
 - **`src/input/`**: Input subsystem (peer of `map/` and `lua/`). `thread.rs` is the producer that blocks on `crossterm::event::read()` and pushes `AppEvent::Input`; `keymap.rs` holds the `KeyMap` table + `KeybindingOverrides` (read from `[keymap]` config and Lua `keymap.set`); `mouse.rs` holds the `MouseAdapter` that translates raw mouse events to `UserIntent`. Keyboard supports count prefixes for pan (e.g., `5j`); `:` opens the command palette, `/` opens the palette pre-loaded with the search provider.
