+++ src/input/thread.rs
@@ -26,7 +26,7 @@ use std::time::Duration;
 use crossterm::event;
 use log::warn;
 
-use crate::frontend::AppEvent;
+use crate::app::AppEvent;
 
 /// Owns the input-reader thread and its shutdown flag.
 ///
