+++ src/lua/op.rs
@@ -1,9 +1,9 @@
-//! Lua → Frontend operation vocabulary.
+//! Lua → App operation vocabulary.
 //!
 //! [`Op`] is the typed output of a Lua subsystem tick / callback: each
 //! variant describes a single same-thread "do this on the App" request
 //! that the Lua bridge enqueues into a shared [`OpsBuffer`] and that
-//! Frontend drains once per iteration.
+//! App drains once per iteration.
 //!
 //! Replaces the older mix of three same-thread mechanisms — the
 //! per-component `CloseFlag` (Arc<AtomicBool> polled every frame),
@@ -15,17 +15,17 @@
 use std::cell::RefCell;
 use std::rc::Rc;
 
+use crate::app::UserIntent;
 use crate::compositor::{CardId, Component};
-use crate::frontend::UserIntent;
 
 /// A same-thread request from the Lua subsystem to the App.
 ///
 /// Lua callbacks (handle `:close()`, `api.card.open`,
 /// `ttymap.map:jump`, …) push these into a shared [`OpsBuffer`];
-/// Frontend drains the buffer once per loop iteration and applies
+/// App drains the buffer once per loop iteration and applies
 /// each op.
 ///
-/// All three Lua → Frontend same-thread variants now ride this one
+/// All three Lua → App same-thread variants now ride this one
 /// buffer (close / push / intent). The earlier mix — `CloseFlag`
 /// polling, per-plugin `push_rx` mpsc, and `LuaSender` mpsc — has
 /// fully retired.
@@ -43,7 +43,7 @@ pub enum Op {
     /// no-op when `id` is not on the stack (handle closed twice, or
     /// the component already self-closed via `win.close()`).
     Close(CardId),
-    /// Dispatch a [`UserIntent`] through `Frontend::dispatch`.
+    /// Dispatch a [`UserIntent`] through `App::dispatch`.
     /// Emitted by Lua-facing host methods (`ttymap.map:jump` /
     /// `:zoom` / `:fly_to`, `ttymap.api.frame.export`, …) — the
     /// canonical intent vocabulary every other emitter (keymap,
@@ -62,12 +62,12 @@ impl std::fmt::Debug for Op {
 }
 
 /// Shared, single-threaded buffer that accumulates [`Op`]s from Lua
-/// callbacks and is drained by Frontend per iteration.
+/// callbacks and is drained by App per iteration.
 ///
 /// `Rc<RefCell<...>>`: same-thread sharing across the API closures
 /// (held inside the Lua VM via captured clones), the returned handles
 /// (`CardHandle` / `PaletteHandle`), and the runtime [`LuaHandle`]
-/// (read by Frontend). The Lua VM is single-threaded by mlua design,
+/// (read by App). The Lua VM is single-threaded by mlua design,
 /// and the buffer never crosses threads.
 pub type OpsBuffer = Rc<RefCell<Vec<Op>>>;
